requirements.txt
black~=24.4.2
flake8~=7.0.0
isort~=5.13.2
pipenv~=2023.12.1
pyautogen~=0.2.27
pycodestyle~=2.11.1
pydocstyle~=6.3.0
pylint~=3.1.0
virtualenv~=20.26.1
yapf~=0.40.2